Rename vendor_assessment package to vetting
Shorter, idiomatic Go name with no underscores.
