+++ pkg/agents/vetting/security_prompt.txt
