+++ pkg/probo/service.go
@@ -24,7 +24,7 @@ import (
 	"go.gearno.de/kit/pg"
 	"go.probo.inc/probo/pkg/agents"
 	"go.probo.inc/probo/pkg/certmanager"
-	vendor_assessment "go.probo.inc/probo/pkg/agents/vendor_assessment"
+	"go.probo.inc/probo/pkg/agents/vetting"
 	"go.probo.inc/probo/pkg/coredata"
 	"go.probo.inc/probo/pkg/crypto/cipher"
 	"go.probo.inc/probo/pkg/esign"
@@ -68,7 +68,7 @@ type (
 		slack                   *slack.Service
 		esign                   *esign.Service
 		invitationTokenValidity time.Duration
-		vendorAssessor         *vendor_assessment.Assessor
+		vendorAssessor         *vetting.Assessor
 	}
 
 	TenantService struct {
@@ -80,7 +80,7 @@ type (
 		baseURL                           string
 		tokenSecret                       string
 		agent                             *agents.Agent
-		vendorAssessor                    *vendor_assessment.Assessor
+		vendorAssessor                    *vetting.Assessor
 		fileManager                       *filemanager.Service
 		esign                             *esign.Service
 		Frameworks                        *FrameworkService
@@ -145,7 +145,7 @@ func NewService(
 	iamService *iam.Service,
 	esignService *esign.Service,
 	invitationTokenValidity time.Duration,
-	vendorAssessor *vendor_assessment.Assessor,
+	vendorAssessor *vetting.Assessor,
 ) (*Service, error) {
 	if bucket == "" {
 		return nil, fmt.Errorf("bucket is required")
